Fixed bugs and implemented minor changes
Fixed calculator disappearing when you press "." when the calculator was empty. When you press "." when the calculator is empty it will instead input "0." and you can either add numbers or not. Also fixed dividing by zero and 0.1.

Added a alert if you try to divide by zero.

You can also use "=" key to evaluate.